Split up Response into two classes for a cleaner implementation.
