+++ includes/classes/AdvancedCache/Data/Response.php
@@ -5,12 +5,10 @@
 
 namespace DarkMatter\AdvancedCache\Data;
 
-use DarkMatter\Interfaces\Registerable;
-
 /**
  * Class Response
  */
-class Response implements Registerable {
+class Response {
 	/**
 	 * Body as string. This is most commonly HTML but can also be JSON, XML (i.e. RSS), or plaintext.
 	 *
@@ -81,30 +79,5 @@ class Response implements Registerable {
 		$entry->headers = $this->headers;
 
 		return $entry->save();
-	}
-
-	/**
-	 * Register hooks.
-	 *
-	 * Note: used for processing a live request within WordPress. Can be ignored otherwise.
-	 *
-	 * @return void
-	 */
-	public function register() {
-		add_filter( 'status_header', [ $this, 'set_status_header' ], 10, 2 );
-	}
-
-	/**
-	 * Set the status header from the WordPress hook.
-	 *
-	 * Note: used for processing a live request within WordPress. Can be ignored otherwise.
-	 *
-	 * @param string  $status_header HTTP status header.
-	 * @param integer $status_code   HTTP status code.
-	 * @return string
-	 */
-	public function set_status_header( $status_header = '', $status_code = 200 ) {
-		$this->status_code = $status_code;
-		return $status_header;
 	}
 }
